+++ .github/workflows/mkdocs-to-gh-pages.yml
@@ -2,7 +2,7 @@ name: mkdocs-to-gh-pages
 on:
   push:
     branches:
-      - 9-allow-addition-of-arbitrary-attributes-to-metadata-profiles
+      - master
 
 permissions:
   contents: write
@@ -25,12 +25,8 @@ jobs:
           python-version: 3.x
 
       - name: Install mkdocs and plugins
-        run: |      
-          pip install mkdocs
-          pip install mkdocs-autorefs
-          pip install mkdocs-material
-          pip install mkdocstrings[python]
-          pip install mkdocs-git-revision-date-localized-plugin
+        run: |
+          pip install .[docs]
 
       - name: Build and push docs
         run: mkdocs gh-deploy --force
